add table labels
