+++ app/views/assignments/_chat.html.erb
@@ -1,4 +1,7 @@
-<h3 class="mt-3"><i class="fas fa-comments me-3"></i>Chatroom
+<h3 class="mt-3">
+  <i class="fas fa-comments me-3"></i>
+  <%= t('resources.chatroom') %>
+
   <span  class="navbar-toggler" type="button" data-bs-toggle="modal" data-bs-target="#myModal">
     <i class="fas fa-bars border p-2 rounded"></i>
   </span>
