+++ app/views/devise/passwords/new.html.erb
@@ -6,7 +6,7 @@
     <div class="form-inputs">
       <%= f.input :email,
                   required: true,
-                  label: t('profile.email'),
+                  label: t('label.user.email'),
                   autofocus: true,
                   input_html: { autocomplete: "email" } %>
     </div>
