+++ app/views/devise/sessions/new.html.erb
@@ -6,12 +6,12 @@
     <div class="form-inputs">
       <%= f.input :email,
                   required: false,
-                  label: t('profile.email'),
+                  label: t('label.user.email'),
                   autofocus: true,
                   input_html: { autocomplete: "email" } %>
       <%= f.input :password,
                   required: false,
-                  label: t('profile.password'),
+                  label: t('label.user.password'),
                   input_html: { autocomplete: "current-password" } %>
       <%= f.input :remember_me, label: t('.checkbox_remember_me'), as: :boolean if devise_mapping.rememberable? %>
     </div>
