+++ app/views/shared/_sidebarMenu.html.erb
@@ -60,7 +60,7 @@
           <%= link_to tutors_path, id: 'tutors-menu', class: "card-link text-dark text-decoration-none" do %>
             <div <%= @active_users %>>
               <i class="fas fa-chalkboard-teacher text-secondary me-1"></i>
-              <%= t('resources.tutors') %>
+              <%= t('resources.tutor') %>
             </div>
           <% end %>
         </h5>
